+++ tests/testthat/test-fastcpd.R
@@ -7,7 +7,7 @@ testthat::test_that(
       ),
       r"[
 The family should be one of "gaussian", "binomial", "poisson", "lasso", "ar",
-"custom" or `NULL` while the provided family is bin0mial.]"
+"var", "custom" or `NULL` while the provided family is bin0mial.]"
     )
   }
 )
@@ -821,6 +821,59 @@ testthat::test_that(
     )
 
     testthat::expect_equal(result@cp_set, 615)
+  }
+)
+
+testthat::test_that(
+  "example var(2) model with p = 2 with innovation sd = 1", {
+    set.seed(1)
+    n <- 1000
+    p <- 2
+    theta_1 <- matrix(c(-0.3, 0.6, -0.5, 0.4, 0.2, 0.1, 0.1, -0.2), nrow = p)
+    theta_2 <- matrix(c(-0.1, -0.2, 0.1, 0.05, 0.1, -0.2, -0.2, 0.1), nrow = p)
+    x <- matrix(0, n + 2, p)
+    for (i in 1:600) {
+      x[i + 2, ] <- theta_1 %*% c(x[i + 1, ], x[i, ]) + rnorm(p, 0, 1)
+    }
+    for (i in 601:1000) {
+      x[i + 2, ] <- theta_2 %*% c(x[i + 1, ], x[i, ]) + rnorm(p, 0, 1)
+    }
+    data <- matrix(NA, n, p * 3)
+    data[, seq_len(p)] <- x[p + seq_len(n), ]
+    for (p_i in seq_len(p)) {
+      data[, p + (p * p_i - 1):(p * p_i)] <- x[p - p_i + seq_len(n), ]
+    }
+    multi_response_linear_loss <- function(data) {
+      x <- data[, (ncol(data) - p + 1):ncol(data)]
+      y <- data[, 1:(ncol(data) - p)]
+
+      if (nrow(data) <= p + 1) {
+        x_t_x <- diag(p)
+      } else {
+        x_t_x <- crossprod(x)
+      }
+
+      norm(y - x %*% solve(x_t_x, t(x)) %*% y, type = "F")^2 / 2
+    }
+    result <- fastcpd(
+      formula = y.1 + y.2 ~ x.1 + x.2 + x.3 + x.4 - 1,
+      data = data.frame(y = data[, 1:2], x = data[, 3:6]),
+      beta = (2 * 4 + 1) * log(n) / 2,
+      p = 4,
+      cost = multi_response_linear_loss
+    )
+
+    testthat::expect_equal(result@cp_set, 612)
+
+    result_var <- fastcpd(
+      formula = ~ . - 1,
+      data = data.frame(x = x),
+      beta = (2 * 4 + 1) * log(n) / 2,
+      p = 2,
+      family = "var"
+    )
+
+    testthat::expect_equal(result_var@cp_set, 611)
   }
 )