+++ NEWS.md
@@ -8,6 +8,7 @@
 *   Separate the use of internal C++ cost functions and user-defined R cost
     functions.
 *   Add Codecov Icicle plot in README.
+*   Remove `cost_optim` and `cost_update` from `RcppExports.R`.
 
 # fastcpd 0.7.6