+++ NEWS.md
@@ -2,6 +2,7 @@
 
 *   Add `fastcpd.ts` / `fastcpd_ts` for time series data.
 *   Fix pre segmengation bug for `lasso`.
+*   Fix bug related to `vanilla_percentage` parameter for `lasso`.
 
 # fastcpd 0.8.2